Changed problems file structure
