+++ openqaoa/problems/__init__.py
@@ -1,0 +1,6 @@
+from .knapsack import Knapsack, SlackFreeKnapsack
+from .maximumcut import MaximumCut
+from .minimumvertexcover import MinimumVertexCover
+from .numberpartition import NumberPartition
+from .shortestpath import ShortestPath
+from .tsp import TSP
